requirements.txt
asgiref==3.3.1
Django==3.1.7
django-colorfield==0.4.1
pkg-resources==0.0.0
pytz==2021.1
sqlparse==0.4.1
